requirements.txt
click==6.6
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
ordereddict==1.1
python-dotenv==0.5.0
redis==2.10.5
redis-orm>=0.4.0
Werkzeug==0.11.9
wrapt==1.10.8
